Moved parser functions and fixed KeyError issue for xml_sivin_parser function
