+++ psa/forms.py
@@ -2,7 +2,8 @@ from django import forms
 from django.utils.translation import gettext as _
 from bootstrap_modal_forms.forms import BSModalModelForm
 
-from utils.django.validators import validate_vin, validate_nac, xml_parser
+from utils.scraping import xml_parser
+from utils.django.validators import validate_vin, validate_nac
 from .models import Corvet, Firmware
